# docs/api-docs/openapi.yaml
openapi: 3.0.3
info:
  title: SignVault API
  version: 0.2.0
  description: |
    Kontrak API SignVault v0.2 (sumber kebenaran, prd.md §10).
    Scope: backend (apps/backend, packages/database, docs/api-docs).
    v0.2: Google OAuth (POST /auth/google), onboarding (PATCH
    /users/me/complete-profile), CAPTCHA Turnstile, rate-limit auth 10/mnt.
    Base path: /api/v1.
servers:
  - url: http://localhost:8000/api/v1
    description: Local dev (Uvicorn native, tanpa Docker)
...
paths:
...
  /keys/generate:
    post:
      tags: [keys]
      summary: Generate key pair (private key dienkripsi AES-256-GCM)
      operationId: generateKey
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateKeyRequest'
      responses:
        '422':
          $ref: '#/components/responses/ValidationError'
        '201':
          description: Key pair terbuat (private key TIDAK pernah dikembalikan plaintext)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KeyPair'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /keys:
    get:
      tags: [keys]
      summary: Daftar key pair milik signer
      operationId: listKeys
      responses:
        '200':
          description: Daftar key
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/KeyPair'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /keys/{id}/revoke:
    post:
      tags: [keys]
      summary: Revoke key (Super Admin). Key rotation = revoke + generate baru
      operationId: revokeKey
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Key di-revoke
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KeyPair'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /documents:
    post:
      tags: [documents]
      summary: Upload PDF (Org Admin, max 25 MB)
      operationId: uploadDocument
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/UploadDocumentRequest'
      responses:
        '422':
          $ref: '#/components/responses/ValidationError'
        '201':
          description: Dokumen terupload, hash SHA-256 dihitung
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    get:
      tags: [documents]
      summary: Daftar dokumen (paginated, filter status)
      operationId: listDocuments
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: status
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/DocumentStatus'
      responses:
        '200':
          description: Daftar dokumen
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentPage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /documents/{id}:
    get:
      tags: [documents]
      summary: Detail dokumen (butuh auth)
      operationId: getDocument
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Detail dokumen
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /documents/{id}/download:
    get:
      tags: [documents]
      summary: Unduh PDF asli / bertanda tangan (privat, butuh auth)
      operationId: downloadDocument
      parameters:
        - $ref: '#/components/parameters/IdParam'
        - name: kind
          in: query
          required: false
          schema:
            type: string
            enum: [original, signed]
            default: original
      responses:
        '200':
          description: File PDF (Content-Disposition inline)
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /documents/{id}/qr-placements:
    put:
      tags: [documents]
      summary: Simpan posisi QR (fraksi 0..1, origin kiri-atas). Pemilik + DRAFT saja
      operationId: saveQrPlacements
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [placements]
              properties:
                placements:
                  type: array
                  maxItems: 10
                  items:
                    type: object
                    required: [page, x, y, size]
                    properties:
                      page:
                        type: integer
                        minimum: 1
                      x:
                        type: number
                        minimum: 0
                        maximum: 1
                      y:
                        type: number
                        minimum: 0
                        maximum: 1
                      size:
                        type: number
                        minimum: 0.02
                        maximum: 0.8
      responses:
        '422':
          $ref: '#/components/responses/ValidationError'
        '200':
          description: Posisi tersimpan
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /documents/{id}/request-sign:
    post:
      tags: [documents]
      summary: Org Admin meminta tanda tangan ke Signer
      operationId: requestSign
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestSignRequest'
      responses:
        '422':
          $ref: '#/components/responses/ValidationError'
        '201':
          description: SignRequest terbuat + notifikasi terkirim ke Signer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SignRequest'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
    PageParam:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  responses:
    BadRequest:
      description: Request tidak valid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
    Unauthorized:
      description: Token hilang / tidak valid / kedaluwarsa
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
    Forbidden:
      description: Role tidak diizinkan
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
    ValidationError:
      description: Validasi request gagal (tipe/format field salah)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
...
    NotFound:
      description: Resource tidak ditemukan
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
    Conflict:
      description: Konflik status (mis. sudah di-sign / double request)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
  schemas:
    ErrorBody:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: UNAUTHORIZED
            message:
              type: string
              example: Token tidak valid atau kedaluwarsa.
...
    KeyAlgorithm:
      type: string
      enum: [RSA_PSS_2048, ECDSA_P256, ED25519]
    DocumentStatus:
      type: string
      enum: [DRAFT, PENDING, SIGNED, REJECTED]
    SignRequestStatus:
      type: string
      enum: [PENDING, APPROVED, REJECTED]
...
    GenerateKeyRequest:
      type: object
      required: [algorithm]
      properties:
        algorithm:
          $ref: '#/components/schemas/KeyAlgorithm'
    KeyPair:
      type: object
      required: [id, algorithm, publicKey, revoked, createdAt]
      properties:
        id:
          type: string
        algorithm:
          $ref: '#/components/schemas/KeyAlgorithm'
        publicKey:
          type: string
          description: PEM / base64 public key
        revoked:
          type: boolean
        revokedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
    UploadDocumentRequest:
      type: object
      required: [file, title]
      properties:
        file:
          type: string
          format: binary
          description: File PDF, max 25 MB
        title:
          type: string
        description:
          type: string
        metadata:
          type: string
          description: >-
            JSON string metadata kanonis yang ikut ditandatangani. Key standar
            form unggah: nomor (maks 50), tanggal (YYYY-MM-DD), jenis (salah satu
            dari 14 jenis dokumen), pengirim (maks 100).
    Document:
      type: object
      required: [id, title, storagePath, fileHash, metadata, status, version, createdAt, updatedAt]
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
          nullable: true
        storagePath:
          type: string
        fileHash:
          type: string
          description: SHA-256 hex dari byte PDF asli
        metadata:
          type: object
        status:
          $ref: '#/components/schemas/DocumentStatus'
        version:
          type: integer
          description: Optimistic locking untuk batch sign
        pageCount:
          type: integer
          description: Jumlah halaman PDF
        qrPlacements:
          type: array
          description: Posisi QR fraksi 0..1 origin kiri-atas, dipakai saat signing
          items:
            type: object
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    RequestSignRequest:
      type: object
      required: [signerId]
      properties:
        signerId:
          type: string
        message:
          type: string
    SignRequest:
      type: object
      required: [id, documentId, signerId, status, createdAt, updatedAt]
      properties:
        id:
          type: string
        documentId:
          type: string
        signerId:
          type: string
        status:
          $ref: '#/components/schemas/SignRequestStatus'
        message:
          type: string
          nullable: true
        rejectReason:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    DocumentPage:
      type: object
      required: [data, page, limit, total]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Document'
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
...
security:
  - bearerAuth: []